refactor(cli): migrate inference-config.js to TypeScript
Convert bin/lib/inference-config.js (143 lines) to src/lib/inference-config.ts
with typed interfaces for ProviderSelectionConfig and GatewayInference.

All 3 exported functions are pure — straightforward full conversion.
Reduces switch/case duplication via shared base object. Co-locates tests.

Relates to #924 (shell consolidation).
